Añadir endpoints y métodos a ServicioUsuarios y a UsuarioControlador

+++ src/main/java/edu/comillas/icai/gitt/pat/spring/practica_final/servicio/ServicioPistas.java
@@ -66,9 +66,9 @@ public class ServicioPistas {
         return ResponseEntity.ok(pista);
     }
 
-    public ResponseEntity<Pista> actualizarPista(int idPista, Pista pistaActualizada) {
+    public ResponseEntity<Pista> actualizarPista(long idPista, Pista pistaActualizada) {
 
-        Pista pista = repoPista.findById((long) idPista).orElse(null);
+        Pista pista = repoPista.findById(idPista).orElse(null);
 
         if (pista == null) {
             throw new ResponseStatusException(
